# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: LMS Finance & Voucher Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /instructors:
    get:
      operationId: listInstructors
      tags: [instructors]
      summary: List all instructors
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of instructors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Instructor"
    post:
      operationId: createInstructor
      tags: [instructors]
      summary: Create a new instructor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstructorInput"
      responses:
        "201":
          description: Instructor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instructor"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /instructors/{id}:
    get:
      operationId: getInstructor
      tags: [instructors]
      summary: Get instructor detail with payment history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Instructor detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructorDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateInstructor
      tags: [instructors]
      summary: Update an instructor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstructorInput"
      responses:
        "200":
          description: Updated instructor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instructor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteInstructor
      tags: [instructors]
      summary: Delete an instructor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /instructor-payments:
    get:
      operationId: listInstructorPayments
      tags: [instructors]
      summary: List instructor salary payments
      parameters:
        - name: instructorId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstructorPayment"
    post:
      operationId: createInstructorPayment
      tags: [instructors]
      summary: Record a salary payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstructorPaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructorPayment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /instructor-payments/{id}:
    delete:
      operationId: deleteInstructorPayment
      tags: [instructors]
      summary: Delete a salary payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]

    # ── Course ────────────────────────────────────────────────────────────────
...
    Instructor:
      type: object
      properties:
        id:
          type: integer
        instructorCode:
          type: string
        name:
          type: string
        fatherName:
          type: string
        phone:
          type: string
        address:
          type: string
        specialization:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        monthlySalary:
          type: number
        joinDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - instructorCode
        - name
        - monthlySalary
        - joinDate
        - status
        - createdAt
    InstructorDetail:
      allOf:
        - $ref: "#/components/schemas/Instructor"
        - type: object
          properties:
            payments:
              type: array
              items:
                $ref: "#/components/schemas/InstructorPayment"
          required: [payments]
    CreateInstructorInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        fatherName:
          type: string
        phone:
          type: string
        address:
          type: string
        specialization:
          type: string
        courseId:
          type: integer
          nullable: true
        monthlySalary:
          type: number
          minimum: 0
        joinDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive]
      required:
        - name
        - monthlySalary
        - joinDate
    InstructorPayment:
      type: object
      properties:
        id:
          type: integer
        paymentNumber:
          type: string
        instructorId:
          type: integer
        instructorName:
          type: string
        instructorCode:
          type: string
        specialization:
          type: string
        courseName:
          type: string
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
        monthName:
          type: string
        amountPaid:
          type: number
        paymentMethod:
          type: string
        paymentDate:
          type: string
          format: date
        remarks:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - paymentNumber
        - instructorId
        - instructorName
        - instructorCode
        - month
        - year
        - monthName
        - amountPaid
        - paymentMethod
        - paymentDate
        - createdAt
    CreateInstructorPaymentInput:
      type: object
      properties:
        instructorId:
          type: integer
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
        amountPaid:
          type: number
          minimum: 0
        paymentMethod:
          type: string
          enum: [cash, bank_transfer, cheque, online]
        paymentDate:
          type: string
          format: date
        remarks:
          type: string
      required:
        - instructorId
        - month
        - year
        - amountPaid
        - paymentMethod
        - paymentDate